_ref_python_modbus/python_modbus-main/requirements.txt
psycopg2
python-dotenv
huawei-solar-tools
